+++ FABRIC-QUILT/1.21/src/main/java/elocindev/item_obliterator/fabric_quilt/ItemObliterator.java
@@ -20,7 +20,7 @@ public class ItemObliterator implements ModInitializer {
 	public static ConfigEntries Config = ConfigEntries.INSTANCE;
 
 	public static Set<String> blacklisted_items;
-	public static Set<String> blacklisted_nbt;
+	public static Set<String> blacklisted_component_data;
 	public static Set<String> only_disable_interactions;
 	public static Set<String> only_disable_attacks;
     public static Set<String> only_disable_recipes;
@@ -47,7 +47,7 @@ public class ItemObliterator implements ModInitializer {
 
 	public static void reloadConfigHashsets() {
 		blacklisted_items = new HashSet<>(ItemObliterator.Config.blacklisted_items);
-		blacklisted_nbt = new HashSet<>(ItemObliterator.Config.blacklisted_nbt);
+		blacklisted_component_data = new HashSet<>(ItemObliterator.Config.blacklisted_component_data);
 		only_disable_interactions = new HashSet<>(ItemObliterator.Config.only_disable_interactions);
 		only_disable_attacks = new HashSet<>(ItemObliterator.Config.only_disable_attacks);
 		only_disable_recipes = new HashSet<>(ItemObliterator.Config.only_disable_recipes);