+++ FABRIC-QUILT/1.21/src/main/java/elocindev/item_obliterator/fabric_quilt/mixin/VillagerTradeMixin.java
@@ -6,10 +6,7 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import elocindev.item_obliterator.fabric_quilt.util.Utils;
-import net.minecraft.nbt.NbtCompound;
-import net.minecraft.nbt.NbtList;
 import net.minecraft.screen.MerchantScreenHandler;
-import net.minecraft.util.Util;
 import net.minecraft.village.TradeOfferList;
 
 @Mixin(MerchantScreenHandler.class)
@@ -18,8 +15,9 @@ public class VillagerTradeMixin {
     public void item_obliterator$removeTrades(CallbackInfoReturnable<TradeOfferList> info) {
         if(info.getReturnValue() != null) {
             
-            TradeOfferList Offers = new TradeOfferList(Util.make(new NbtCompound(), 
-                tag -> tag.put("Recipes", new NbtList())));
+            // TradeOfferList Offers = new TradeOfferList(Util.make(new NbtCompound(), 
+            //     tag -> tag.put("Recipes", new NbtList())));
+            TradeOfferList Offers = new TradeOfferList();
 
             info.getReturnValue().forEach(offer -> {
                 if(!Utils.isDisabled(offer.getSellItem()))