+++ FABRIC-QUILT/1.21/src/main/java/elocindev/item_obliterator/fabric_quilt/util/Utils.java
@@ -2,10 +2,10 @@ package elocindev.item_obliterator.fabric_quilt.util;
 
 import dev.emi.emi.api.stack.EmiStack;
 import elocindev.item_obliterator.fabric_quilt.ItemObliterator;
+import net.minecraft.component.Component;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
-import net.minecraft.nbt.NbtCompound;
 import net.minecraft.registry.Registries;
 
 public class Utils {
@@ -14,13 +14,30 @@ public class Utils {
     }  
 
     public static boolean shouldRecipeBeDisabled(Item item) {
-        return isDisabled(getItemId(item))
-        || ItemObliterator.Config.only_disable_recipes.contains(getItemId(item));
+        return shouldRecipeBeDisabled(getItemId(item));
     }
 
     public static boolean shouldRecipeBeDisabled(String itemid) {
-        return isDisabled(itemid)
-        || ItemObliterator.Config.only_disable_recipes.contains(itemid);
+        if (isDisabled(itemid)) return true;
+
+        if (!ItemObliterator.Config.use_hashmap_optimizations) {
+            for (String blacklisted_id : ItemObliterator.Config.only_disable_recipes) {
+                if (blacklisted_id == null) continue;
+                if (blacklisted_id.startsWith("//")) continue;
+                
+                if (blacklisted_id.equals(itemid)) return true;
+
+                if (blacklisted_id.startsWith("!")) {
+                    blacklisted_id = blacklisted_id.substring(1);
+
+                    if (itemid.matches(blacklisted_id)) return true;
+                }
+            }
+        } else {
+            if (ItemObliterator.only_disable_recipes.contains(itemid)) return true;
+        }
+
+        return false;
     }
 
     public static boolean isDisabled(String itemid) {
@@ -51,7 +68,9 @@ public class Utils {
     public static boolean isDisabled(EmiStack emiStack) {
         if (emiStack == null) return false;
     
-        if (emiStack.hasNbt() && isDisabled(emiStack.getNbt())) return true;
+        // bye bye nbt!!
+        // TODO: reimplement this whenever emi updates
+        //if (emiStack.hasNbt() && isDisabled(emiStack.getNbt())) return true;
     
         if (emiStack.getKey() instanceof Item item) {
             return isDisabled(getItemId(item));
@@ -64,28 +83,28 @@ public class Utils {
     public static boolean isDisabled(ItemStack stack) {
         if (stack == null || stack.isOf(Items.AIR)) return false;
 
-        if (stack.hasNbt()) {
-            isDisabled(stack.getNbt());
+        for (var component : stack.getComponents()) {
+            isDisabled(component);
         }
         
         return isDisabled(getItemId(stack.getItem()));
     }
 
-    public static boolean isDisabled(NbtCompound nbt) {
-        if (nbt == null) return false;
+    public static boolean isDisabled(Component<?> component) {
+        if (component == null) return false;
 
-        for (String blacklisted_nbt : ItemObliterator.Config.blacklisted_nbt) {
-            if (blacklisted_nbt == null) continue;
-            if (blacklisted_nbt.startsWith("//")) continue;
+        for (String blacklisted_data : ItemObliterator.Config.blacklisted_component_data) {
+            if (blacklisted_data == null) continue;
+            if (blacklisted_data.startsWith("//")) continue;
 
-            String nbtString = nbt.toString();
+            String componentString = component.toString();
 
-            if (nbtString.contains(blacklisted_nbt)) return true;
+            if (componentString.contains(blacklisted_data)) return true;
 
-            if (blacklisted_nbt.startsWith("!")) {
-                blacklisted_nbt = blacklisted_nbt.substring(1);
+            if (blacklisted_data.startsWith("!")) {
+                blacklisted_data = blacklisted_data.substring(1);
 
-                if (nbtString.matches(blacklisted_nbt)) return true;
+                if (componentString.matches(blacklisted_data)) return true;
             }
         }